backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
websockets==14.1
websockets_proxy==0.1.3
elevenlabs>=1.2.1
numpy>=1.26.0
